[turbofan] Emit WrongValue-deopt checks for const stores in literals
When storing to a const property we typically only emit StoreField for the initial store; consecutive stores are translated to a load of the existing value and compared against the new value that is about to be stored. If the two don't match, we deopt with WrongValue, causing the runtime to generalize the property constness to mutable.

So far, this was only done for regular property stores (those of access mode kStore), whereas stores that are part of literal initialization (those of access mode kStoreInLiteral) would simply emit additional stores. By itself, this appears to be sound, since earlier stores to literals cannot be observed anyways.

This CL changes the behavior to always emit the WrongValue-deopt check, i.e., even in case of kStoreInLiteral. This allows us to maintain the double const store invariant in load elimination (i.e., we may never observe two consecutive const stores to the same property on the same object). It also aligns turbofan with the runtime behavior, which always checks for same values on later const stores and generalizes when they don't agree.

R=ishell@chromium.org

Bug: chromium:987205
Change-Id: I974028cc13ba57b81b8fdd3c349650770500ec37