Update patch set 9
Patch Set 9:

> Patch Set 8:
> 
> (6 comments)
> 
> Looks good overall. Could you add a couple of test-bytecode-generator tests please?

It seems the lazily compiled functions won't show up in the bytecode expectations - since the private members have to be accessed through some public method, the code of access (in the public member) is always lazily generated. Wrapping the evaluation of the class in an expression does not seem to do the trick either. Is there a way to get around that?

Patch-set: 9
Reviewer: Gerrit User 1288325 <1288325@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>