Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > > Patch Set 8:
> > > 
> > > Premonomorphic is still used for global stores. Is that the only remaining use?
> > 
> > Yes, as far as I know that is the only use.
> 
> Can we remove it?

When a global property doesn't exist we run into a interceptor which then installs the property. When we run into a interceptor we just install a slow stub and always stay with the slow stub. To avoid this pre-monomorphic state was introduced recently for global stores (https://bugs.chromium.org/p/v8/issues/detail?id=8712). We decided to leave this as is for now. I am not really sure how often we hit this case even with lazy feedback allocation.

Patch-set: 8