Update patch set 7
Patch Set 7:

Sorry for the slow response, followed up on the filed bug with more details, but I think this expression is incorrect in the overview document. I think the right fix is to update the overview document to use only the lower bits instead of a mod operation on the shift value.

Patch-set: 7
Reviewer: Gerrit User 1122079 <1122079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>