Reduce the number of ints tested
This reduces the runtime from ~20m to ~2m (very unscientific measure
based on running the entire asm-wasm-i32 test with and without this
change).

I removed most of the constants that looked uninteresting, e.g. testing
for 10, 20, 30, isn't that interesting. The edge cases are left
untouched, max singed positive/negative ints and +/- 1 from both.

Bug: v8:7783 v8:9396
Change-Id: Ice363fc3f786dd55ff118ffa42f9ecea07880338