Update patch set 3
Patch Set 3: Code-Review+1

Thanks for implementing this set of operations! As it wasn't always clear cut whether this should be included in v1, once this lands it may also be a good idea to verify that both sets of operations do get used in the benchmark cited in the corresponding SIMD issue - https://github.com/WebAssembly/simd/issues/13. 

I've verified that the code invokes both floats and doubles, but would also be good to include this in the benchmarking directory for doubles if we don't already have a benchmark that uses sqrts.

Patch-set: 3
Reviewer: Gerrit User 1122079 <1122079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1