Update patch set 3
Patch Set 3:

This seems like overkill to me: If we simply mark the variable with is_static we'll have access to this everywhere and we don't need to duplicate this information into AssignType and VariableMode? We should immediately be able to detect conflicts based on mismatching static flag when the parser tries to complement accessors.

Let's try to keep independent bits of information separate for simplicity.

Patch-set: 3
Reviewer: Gerrit User 1184114 <1184114@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>