Update patch set 18
Patch Set 18:

I kinda like the proposal to track positions for "this" references listed in the doc:

You can basically consider "this" to be a variable with an initializer that "finishes" when the first top-level super() appears (as approximation). Any reference before the initializer is a reference that requires a hole check. Anything after super() can elide the hole check.

You could even simply have 2 ThisExpression() objects, ThisExpressionWithHoleCheck() and ThisExpression() depending on whether you are before this super() call or not.

Patch-set: 18
Reviewer: Gerrit User 1184114 <1184114@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>