Update patch set 23
Patch Set 23:

> Patch Set 23:
> 
> (3 comments)
> 
> Thanks for the reviews! Hopefully this was actually worth doing at which point we can think about implementing it with source positions on the ThisExpression.

The current solution actually looks nice to me: It seems pretty likely that super() will be a top-level expression.

I'm actually not sure it's easily possible to further optimize without doing control-flow-analysis:

if (...) {
  super()
} else {
  // omit super
}
this // needs a hole check

if (...) {
  super()
} else {
  super()
}
this // doesn't need a hole check

while (...) {
  super()
}
this // needs a whole check unless the loop is guaranteed to run

try {
  throw;
  super
} catch(e) {}
this // needs hole check

etc.

The simple cases are already dealt with in this CL. The reason why source position works for let is because the initializer is guaranteed to finish in a dominator block of any uses, but that's not the case for super() unfortunately.

Patch-set: 23
Reviewer: Gerrit User 1184114 <1184114@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>