Update patch set 8
Patch Set 8:

update: it seems that we currently always expect a named class variable to be used since it's declared as a bound variable:

https://cs.chromium.org/chromium/src/v8/src/parsing/parser.cc?q=declareclassv&sq=package:chromium&g=0&l=2768

And there is a DCHECK here that raises if the class variable is unused and unallocated:

https://cs.chromium.org/chromium/src/v8/src/interpreter/bytecode-generator.cc?q=BuildClassLiteral&sq=package:chromium&g=0&l=2164

There is also a DCHECK that raises if set the variable to be used and has been forced context allocation at the same time.

https://cs.chromium.org/chromium/src/v8/src/ast/scopes.cc?q=mustallocate&sq=package:chromium&g=0&l=2099

I think to implement the static brand checking via loading the class variable from context, we have two options:

1. Always allocate the class variable in the context, if itself or any private static methods are used in inner scopes
2. Allocate the class variable in the stack if only itself referenced in inner scopes. Allocate it in the context if it's transitively referenced from any private static methods.

I think Toon's suggestion was leaning towards but we cannot use ForceContextAllocation because of the assertion in MustAllocate. Should we just lift that DECHECK? Or make an exception for class variables there?

Patch-set: 8
Reviewer: Gerrit User 1288325 <1288325@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0