Mark functions for optimization only on bytecode budget interrupts
We used to mark functions for optimization on any interrupt. This sometimes
causes functions to OSR when not needed. The implementation was such because
we didn't have a different runtime function to distinguish bytecode budget
interrupts from other interrupts. For lazy feedback allocation we added a
new runtime function for bytecode budget interrupts so it makes it easier
to actually mark functions only when needed.

This also includes a fix to the StackCheck handler to check for terminate
exception flag.

Bug:chromium:993061
Change-Id: Iaf7b4dccb7a503e5b6bfcbb993bc7482aa593955