Update patch set 2
Patch Set 2:

FYI: On the webrtc CL to disable goma, we got confused a bit because goma implicitly sets symbol_level=1 -- you very likely want to explicitly set that to keep the bots fast(ish) and to not build with full symbols all of a sudden.

(But maybe you do that already. With webrtc we ran into mspdbsrv.exe-related issues, while this change here landed fine.)

Patch-set: 2
CC: Gerrit User 1001629 <1001629@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>