Update patch set 2
Patch Set 2:

Can we reconsider this? We tried a solution using DevTools but the runtime overhead was too high since it requires the Debugger domain (for more information see https://github.com/nodejs/diagnostics/issues/320). Even if the overhead was acceptable, the Debugger domain is not recommended for production, which is the main use case for postmortem tools.

Patch-set: 2
CC: Gerrit User 1344620 <1344620@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>