[ptr-compr] Implement the DepressionAvoider Reducer and its phase
The DepressionAvoider aims to avoid adding the root in AnyTagged or
TaggedPointer loads. For the TaggedSigned case, we already solve it
in instruction selection.

The new phase will run only when pointer compression is enabled. For the
moment, it's also requires FLAG_turbo_decompression_elimination to be
false. This latter flag is only temporary to test out the implementation.

Bug: v8:7703
Change-Id: I6b4aec203ab8cbb540b2513cabb1e2a5691ce938