Update patch set 6
Patch Set 6:

(5 comments)

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > (3 comments)
> > 
> > I'm going to comment on a different approach for arm64 on the other review tomorrow (https://chromium-review.googlesource.com/c/v8/v8/+/1859618)
> 
> I forgot to mention that the 'slow path' doesn't have any tests, is it possible to add one?

There are actually tests that execute the slow path (in the wasm spec test suite), but it's really tricky to test every aspect of it. We would need to ensure that all registers in use keep their value, that all stack slots are correctly cleared to zero, and that no other stack slots are cleared.
This requires some complex setup and relies on interaction between compiler, register allocator and assembler. Apart from the bigger tests we already have (which found bugs in earlier patch sets, but do not exhaustively test every aspect), we usually rely on the fuzzers to find the subtle cornercases.
This is not a perfect solution of course, but it worked reasonable well in the past.

Patch-set: 6
Label: Commit-Queue=0