[Liftoff] Optimize function prologue for arm64
Instead of spilling x0 and x1 to the stack, use temp registers which
should always be available in this situation. This avoids the spill and
load.
Because we now take the two available scratch registers, we need to
avoid the {TurboAssembler::Sub} method, which might need another
scratch register. Instead, do something equivalent manually.

R=joey.gouly@arm.com

Bug: v8:9830
Change-Id: I5fc8b8603afaf12fad2ba66cd7a75555ef31bd4b