Update patch set 1
Patch Set 1:

This was surfaced by running the array-concat test with %ResetAllProtectors. The `obj` object in `testConcatArrayLikeStringLength` has the same shape as `obj` in `testConcatArrayLike`. The isConcatSpreadable keyed store was going through the "existing transition handler" case in the megamorphic keyed store builtin. That case doesn't have a protector check like all the other cases in that builtin.

Patch-set: 1