Update patch set 2
Patch Set 2:

I share the concern that simply resetting protectors leaves the VM in an inconsistent (i.e. dangerous, and also test-only, never-occurring-in-production) state. It may be fine for this test right now, but I'm wary of introducing helpers that are easy to misuse.

I think the safer alternative is to reset the entire VM, or as much of it as necessary. Two approaches come to mind:
(1) Most protectors are per-native-context. You can use d8's existing Realms API to run each of these tests in a new context.
(2) Just turn it off and on again: split this one massive test (which is bad practice for parallelization purposes anyway) into individual files, then each will run in its own process with a fresh VM with unpolluted protectors.

Patch-set: 2
Reviewer: Gerrit User 1115935 <1115935@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>