Update patch set 1
Patch Set 1:

Here is another way to do it (and let CSS do the width adjustments):

Add another wrapper div (a sibling of toolbox-anchor) in the multiview after the toolbox anchor (in the constructor of multiview), and attach the {Graph,Schedule,Sequence}View to this div.

 - remove CSS styling for toolbox-anchor to make it have the right height
 - use position: fixed for graph-toolbox (optional but preferred)
 - remove the margin from #schedule and friends
 - add  display: flex; and  flex-direction: column; to .viewpane styles

This should make it work.

Patch-set: 1
Reviewer: Gerrit User 1119600 <1119600@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>