Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Here is another way to do it (and let CSS do the width adjustments):
> 
> Add another wrapper div (a sibling of toolbox-anchor) in the multiview after the toolbox anchor (in the constructor of multiview), and attach the {Graph,Schedule,Sequence}View to this div.
> 
>  - remove CSS styling for toolbox-anchor to make it have the right height
>  - use position: fixed for graph-toolbox (optional but preferred)
>  - remove the margin from #schedule and friends
>  - add  display: flex; and  flex-direction: column; to .viewpane styles
> 
> This should make it work.

It didn't work.

Originally, I tried doing it in only CSS but I couldn't make it work. We need "position: fixed" for graph-toolbox so it doesn't scroll when the pane scrolls.
However, if we have position as fixed then we would need to update graph-toolbox width. Otherwise, it would have fixed width too: it won't cover the full pane and go over the Disassembly panel to the right.
In essence, we want graph-toolbox width to be equal to be as big as possible so that it covers the whole pane, but not counting the scroll bar (if any). Then, updating that width is that part that's tricky.

I even tried fiddling with the toolbox-anchor width to then maybe inherit that width, but I couldn't find a good solution.

Searching online, the suggestions regarding width and "position: fixed" are to use JS and that this can't be done in pure CSS: it cannot inherit its parent's width since the element no longer has any relation to the parent element.

Patch-set: 1
Reviewer: Gerrit User 1327869 <1327869@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0