[protectors] Add %ResetAllProtectors for testing
Add a testing native to reset the state of all protectors and use it in
the es6/array-concat test.

Since protectors are invalidated once and stay invalidated for the
lifetime of the NativeContext/Isolate, the array-concat test was not
exercising any update protector code paths after the first invalidation
in the test file.

Bug: v8:9837
Change-Id: I8f744b0263c82f1dae61a55032124d9129f8e6f3