[ptr-compr] Implement the DecompressionAvoider Reducer and its phase
The DecompressionAvoider aims to avoid adding the root in AnyTagged or
TaggedPointer loads. For the TaggedSigned case, we already solve it
in instruction selection.

The new phase will run only when pointer compression is enabled. For the
moment, it's also requires FLAG_turbo_decompression_elimination to be
false. This latter flag is only temporary to test out the implementation.

The phase needs to be run when Machine nodes appear in the graph, i.e at the
very end of the pipeline since it them to be present. Also, since this phase
may change the load's MachineRepresentation from Tagged to Compressed, it's
best to run it as late as possible in order to keep the phases that
know about Compressed representation to a minimum.

As an example, if we Load a Tagged value only to Store it back again (i.e
Load -> Store nodes, with the Load being the Store's value) we don't need to
fully decompress it since the Store will ignore the top bits.

Bug: v8:7703
Change-Id: I6b4aec203ab8cbb540b2513cabb1e2a5691ce938