Merged: [Liftoff] Fix out of bounds read in lookahead
The lookahead did not check whether there is actually a byte left to be
read. So if the i32 comparison was the last byte in the function body,
we would read out of memory.
This CL fixes that by introducing a separate {lookahead} method which
does the proper bounds check and the lookahead.

R=​jkummerow@chromium.org

Bug: chromium:1014834, v8:9873
Originally-reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/1864939
No-Try: true
No-Presubmit: true
No-Tree-Checks: true
(cherry picked from commit 1d6542bb65dc502b58fcde8e0d1b8108f95eb849)
Change-Id: Id8b6f74e80eb5eb935f98157a583438495139a2d