[heap] Add Uninitialized wrapper for new allocations
WIP: DO NOT COMMIT

Make heap allocation methods return a new Uninitialized wrapper object,
which wraps the HeapObject and, in debug mode, a DisallowHeapAllocation
scope. We can access the underlying object via operator-> to initialize
it, but can only access it as a HeapObject object via a publish() method,
which also closes the DisallowHeapAllocation scope.

This way we are explicit about when objects returned by the heap are
fully initialized, and ready to be seen by the GC. Additionally, we will
be able to add Foo::Init methods to HeapObject which do the required
initialization, while being explicit about them operating on an
uninitialized object.

Change-Id: I8793e58b4ae35c86053a2fd1df2325a9b249b171