[prettyprinter] Do not try to allocate in case of encoding error
Currently, vsnprintf returns -1, in case of error. Either if the buffer
was not large enough. Or if there was a conversion error.

As a result, in the case of conversion error, the code in
AstPrinter::Print try to allocate more space. And then tries again, and
fails again.

We want to be able to disociate the two kind of errors. In order to
achieve that, VSNPrintF will return -1 in case of an encoding error, and
return the number of characters that would have been printed. It's now
up to the caller to check that the buffer was large enough.

Bug:v8:8934
Change-Id: I5043264de64baf04fd9a7319e2b2b7e8952a88f1