[runtime] Creates a global/read-only ScopeInfo for NativeContext
The native context used an empty function scope info. This is inconsistent with
the fact the native context has an extension slot, since the empty function scope
info doesn't have the extension slot flag set.

This CL creates a scope info dedicated for the native context
with the flag set.

Bug: v8:9744
Change-Id: I00459e9a0ca75dd7a0e2add5e9e61747d0635f39