[torque] Capture owning pointers in std::unique_ptr immediately
In some places objects where allocated on the heap and stored
in a std::unique_ptr later. This CL changes this so that a

std::unique_ptr takes ownership of new objects immediately.
The std::unique_ptr<T>(new T(...)) construct is replaced with
std::make_unique<T>(...) where possible.
Change-Id: Icdb4c9e7536d2b437df1a5bb6c3ad94c97e1e4cc